Migrate function.
